% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank-marketing.R
\docType{data}
\name{bankmkt_smpl}
\alias{bankmkt_smpl}
\title{Portuguese Bank Direct Marketing Campaign}
\format{A data frame with fields:
\describe{
  \item{age}{age (years)}
  \item{job}{type of job}
  \item{marital}{marital status}
  \item{education}{level of education}
  \item{default}{has credit in default?}
  \item{housing}{has a housing loan?}
  \item{loan}{has a personal loan?}
  \item{contact}{mode of previous campaign contact}
  \item{month}{month of previous campaign contact}
  \item{dow}{day of week of previous campaign contact}
  \item{duration}{duration of previous campaign contact (seconds)}
  \item{campaign}{number of contacts during campaign}
  \item{pdays}{number of days since client was last contacted from a previous campaign}
  \item{previous}{number of contacts prior to this campaign}
  \item{poutcome}{outcome of the previous marketing campaign}
  \item{employment}{employment variation rate}
  \item{cpix}{consumer price index}
  \item{ccix}{consumer confidence index}
  \item{euribor}{EURIBOR rate}
  \item{nemployed}{number of citizens employed (thousands)}
  \item{subscribed}{did target subscribe?}
}}
\usage{
bankmkt_smpl
}
\description{
The campaign objective was to persuade targets to make a fixed term deposit.
}
\details{
Data originate from from https://archive.ics.uci.edu/ml/datasets/Bank+Marketing.

The attributes employment, cpix, ccix, euribor and nemployed give information on social and economic context.

There are two data sets:

- bankmkt_full: all data
- bankmkt_smpl: a 10\% random sample

There might be multiple fields for the same target, reflecting multile contacts.
}
\keyword{datasets}
