% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\docType{data}
\name{circle}
\alias{circle}
\title{A circle of red points in a plane of blue points}
\format{A data frame with 100 records and 2 variables:
\describe{
  \item{x}{x coordinate;}
  \item{y}{y coordinate;}
  \item{colour}{either "red" or "blue".}
}}
\usage{
circle
}
\description{
A circle of red points in a plane of blue points
}
\keyword{datasets}
