% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\docType{data}
\name{exam}
\alias{exam}
\title{Hours of study related to passing an exam}
\format{A data frame with 100 records and 2 variables:
\describe{
  \item{hours}{hours spent studying;}
  \item{pass}{pass (1) or fail (0).}
}}
\usage{
exam
}
\description{
A group of 100 students spend between 0 and 12 hours studying for an exam.
How does the number of hours spent studying affect the probability that the
student will pass?
}
\keyword{datasets}
